+++ CHANGES.md
@@ -1,4 +1,4 @@
-# 0.12 - 202?-??-??
+# 0.12 - 2022-12-17
 - Load release group image for releases without cover art
 - Load Wikidata data from release group if missing in release
 - Render track counts on releases with thousands separator